requirements.txt
python-telegram-bot[job-queue]==20.*
google-generativeai>=0.3.0
pydub
gTTS
edge-tts
python-dotenv
ffmpeg-python
supabase>=2.0.0
flask>=3.0.0
